check owner all documents CUD

+++ back/api/deleteLecture.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { deleteLecture } from "../controller/lecture";
+import { checkOwner, deleteLecture } from "../controller/lecture";
 import { responseHandler } from "../lib/common";
 
 const deleteLectureRoute = express();
@@ -10,6 +10,7 @@ deleteLectureRoute.post(
     const { lecture } = req.body;
     console.log(req.body);
     // 유효성 검증 , id..
+    await checkOwner({ lecture, owner: req.user._id });
     return await deleteLecture({ lecture });
   })
 );
