+++ back/api/deletePractice.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { deletePractice } from "../controller/practice";
+import { checkOwner, deletePractice } from "../controller/practice";
 import { responseHandler } from "../lib/common";
 
 const deletePracticeRoute = express();
@@ -10,7 +10,7 @@ deletePracticeRoute.post(
     const { practice } = req.body;
 
     // 유효성 검증 , id..
-
+    await checkOwner({ practice, owner: req.user._id });
     return await deletePractice({ practice });
   })
 );
