+++ back/api/deleteProblem.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { deleteProblem } from "../controller/problem";
+import { checkOwner, deleteProblem } from "../controller/problem";
 import { responseHandler } from "../lib/common";
 
 const deleteProblemRoute = express();
@@ -10,7 +10,7 @@ deleteProblemRoute.post(
     const { problem } = req.body;
 
     // 유효성 검증 , id.. problem_type
-
+    await checkOwner({ problem, owner: req.user._id });
     return await deleteProblem({
       problem,
     });
