+++ back/api/deleteTestcase.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { deleteTestcase } from "../controller/testcase";
+import { checkOwner, deleteTestcase } from "../controller/testcase";
 import { responseHandler } from "../lib/common";
 
 const deleteTestcaseRoute = express();
@@ -11,6 +11,7 @@ deleteTestcaseRoute.post(
 
     // 유효성 검증 , id.. problem_type
 
+    await checkOwner({ testcase, owner: req.user._id });
     return await deleteTestcase({
       testcase,
     });
